Moved UserProfile to Component
