--- src/Pages/UserProfilePage/UserProfilePage.css
@@ -1,62 +1,0 @@
-.user-page-container {
-  margin-top: 40px;
-  display: grid;
-  grid-column-gap: 40px;
-  grid-row-gap: 0px;
-  grid-template-rows: min-content 1fr;
-  background-color: lightgrey;
-  border-radius: 10px;
-}
-.user-name-container {
-  background-color: #323232;
-  font-weight: 400;
-  border-radius: 10px;
-  margin: auto;
-  margin-left: 5%;
-  color: white;
-  padding: 0px 25px;
-  z-index: 99;
-  position: relative;
-  top: 45px;
-}
-
-.user-name-container > h1 {
-  font-weight: 400;
-}
-
-.user-metadata-container {
-  background-color: white;
-  border-radius: 10px;
-}
-
-.user-metadata-content {
-  margin: 5%;
-  font-size: large;
-  font-weight: 400;
-}
-.user-metadata-description {
-  margin-top: 100px;
-  margin-bottom: 40px;
-  padding-left: 10px;
-}
-
-.user-video-container {
-  margin-top: 20px;
-  padding: 15px;
-  padding-top: 15px;
-  background-color: var(--recomendation-bar-color);
-  display: grid;
-  column-gap: 15px;
-  grid-auto-columns: 200px;
-  justify-content: center;
-  grid-template-columns: repeat(
-    auto-fill,
-    minmax(var(--video-preview-width), 1fr)
-  );
-}
-
-@media screen and (max-width: 1300px) {
-  .video-page-container {
-    padding-top: 10px;
-  }
-}
